# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "aarde API",
    "version": "dev",
    "description": "Read-only imagery catalog API served by aarde. Imports are local CLI operations and are not exposed here.\n\nUse **Authorize** to enter the web token from the private token file named in the server startup log. Enter only the token, without the Bearer prefix. Swagger stores authorization in memory and clears it on reload.\n\nAll operations require authentication by default. With AARDE_WEB_PUBLIC_READ=true, registered reads including POST /api/v1/imagery/search accept anonymous requests; this served specification reflects that setting. Explicitly supplied invalid tokens are always rejected. API responses are not cached.\n\nThis reference and /openapi.json are public and contain no catalog data or credentials. Try it out sends real requests to this server.\n\nViewer manifest and layer content always require bearer authentication, even when catalog browsing is public."
  },
  "servers": [
    {
      "url": "/",
      "description": "This aarde server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/v1/info": {
      "get": {
        "operationId": "getInfo",
        "summary": "Get build and access information",
        "tags": [
          "Server"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "description": "Available independently of database health. Authentication state describes this request."
      },
      "head": {
        "operationId": "getInfoHead",
        "summary": "Get build and access information (headers only)",
        "tags": [
          "Server"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "A valid aarde web token is required"
          }
        },
        "description": "Available independently of database health. Authentication state describes this request."
      }
    },
...
    "/api/v1/catalogs": {
      "get": {
        "operationId": "listCatalogs",
        "summary": "List catalogs containing imagery",
        "tags": [
          "Catalog"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalogs"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "head": {
        "operationId": "listCatalogsHead",
        "summary": "List catalogs containing imagery (headers only)",
        "tags": [
          "Catalog"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "A valid aarde web token is required"
          },
          "500": {
            "description": "catalog request failed"
          }
        }
      }
    },
    "/api/v1/imagery": {
      "get": {
        "operationId": "listImagery",
        "summary": "List or filter imagery",
        "tags": [
          "Catalog"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "description": "Exact, case-sensitive IDs. Results are ordered by import time descending, then UUID. Concurrent imports can shift offset-based pages.",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Exact catalog ID. Omit or use an empty string for all catalogs."
          },
          {
            "name": "image_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "minLength": 1,
                "maxLength": 255,
                "description": "Exact, case-sensitive ID: 1\u2013255 UTF-8 bytes, no slashes, backslashes, control characters, or surrounding whitespace."
              },
              "maxItems": 100
            },
            "description": "Repeat this query parameter for up to 100 exact image IDs.",
            "style": "form",
            "explode": true
          },
          {
            "name": "cloud_cover_lt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 100
            },
            "description": "Reported scene cloud cover strictly less than this finite percentage (0\u2013100). Zero matches no known values; 100 excludes fully cloudy scenes. Cannot be combined with cloud_cover_lte. Unknown values are excluded by default when a threshold is supplied; cloud_cover_unknown overrides this. All filters combine with AND before ordering and pagination. Omit for no threshold. An empty query value is invalid.",
            "example": 20
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "description": "Page size; values outside 1\u2013200 are rejected."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            },
            "description": "Number of records to skip."
          },
          {
            "name": "cloud_cover_lte",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 100
            },
            "description": "Reported scene cloud cover at most this finite percentage (0\u2013100). Includes the exact boundary, including 0 and 100. Cannot be combined with cloud_cover_lt. Unknown values are excluded by default; use cloud_cover_unknown=include to retain them. Describes the whole scene, not clouds within the search area. An empty query value is invalid."
          },
          {
            "name": "cloud_cover_unknown",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "exclude",
                "include",
                "only"
              ]
            },
            "description": "Unknown cloud-cover policy. Defaults to exclude with either threshold and include otherwise. exclude without a threshold returns only known values; only returns unknown values and cannot be combined with either threshold. An empty query value is invalid."
          },
          {
            "name": "acquired_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Inclusive acquisition timestamp (RFC3339). A date bound excludes imagery with unknown acquisition time. Must precede acquired_before when both are supplied. An empty query value is invalid."
          },
          {
            "name": "acquired_before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Exclusive acquisition timestamp (RFC3339). For an inclusive Through day, send midnight UTC of the following day. A date bound excludes imagery with unknown acquisition time. An empty query value is invalid."
          }
        ]
      },
      "head": {
        "operationId": "listImageryHead",
        "summary": "List or filter imagery (headers only)",
        "tags": [
          "Catalog"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "A valid aarde web token is required"
          },
          "400": {
            "description": "Invalid pagination, IDs, date bounds, cloud thresholds, or unknown-cloud policy."
          },
          "500": {
            "description": "catalog request failed"
          }
        },
        "description": "Exact, case-sensitive IDs. Results are ordered by import time descending, then UUID. Concurrent imports can shift offset-based pages.",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Exact catalog ID. Omit or use an empty string for all catalogs."
          },
          {
            "name": "image_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "minLength": 1,
                "maxLength": 255,
                "description": "Exact, case-sensitive ID: 1\u2013255 UTF-8 bytes, no slashes, backslashes, control characters, or surrounding whitespace."
              },
              "maxItems": 100
            },
            "description": "Repeat this query parameter for up to 100 exact image IDs.",
            "style": "form",
            "explode": true
          },
          {
            "name": "cloud_cover_lt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 100
            },
            "description": "Reported scene cloud cover strictly less than this finite percentage (0\u2013100). Zero matches no known values; 100 excludes fully cloudy scenes. Cannot be combined with cloud_cover_lte. Unknown values are excluded by default when a threshold is supplied; cloud_cover_unknown overrides this. All filters combine with AND before ordering and pagination. Omit for no threshold. An empty query value is invalid.",
            "example": 20
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "description": "Page size; values outside 1\u2013200 are rejected."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            },
            "description": "Number of records to skip."
          },
          {
            "name": "cloud_cover_lte",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 100
            },
            "description": "Reported scene cloud cover at most this finite percentage (0\u2013100). Includes the exact boundary, including 0 and 100. Cannot be combined with cloud_cover_lt. Unknown values are excluded by default; use cloud_cover_unknown=include to retain them. Describes the whole scene, not clouds within the search area. An empty query value is invalid."
          },
          {
            "name": "cloud_cover_unknown",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "exclude",
                "include",
                "only"
              ]
            },
            "description": "Unknown cloud-cover policy. Defaults to exclude with either threshold and include otherwise. exclude without a threshold returns only known values; only returns unknown values and cannot be combined with either threshold. An empty query value is invalid."
          },
          {
            "name": "acquired_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Inclusive acquisition timestamp (RFC3339). A date bound excludes imagery with unknown acquisition time. Must precede acquired_before when both are supplied. An empty query value is invalid."
          },
          {
            "name": "acquired_before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Exclusive acquisition timestamp (RFC3339). For an inclusive Through day, send midnight UTC of the following day. A date bound excludes imagery with unknown acquisition time. An empty query value is invalid."
          }
        ]
      }
    },
    "/api/v1/imagery/{catalogID}/{imageID}": {
      "get": {
        "operationId": "getImagery",
        "summary": "Get an exact imagery record",
        "tags": [
          "Catalog"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Imagery"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "parameters": [
          {
            "name": "catalogID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Exact, case-sensitive ID: 1\u2013255 UTF-8 bytes, no slashes, backslashes, control characters, or surrounding whitespace."
            },
            "description": "Catalog identifier."
          },
          {
            "name": "imageID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Exact, case-sensitive ID: 1\u2013255 UTF-8 bytes, no slashes, backslashes, control characters, or surrounding whitespace."
            },
            "description": "Image identifier within the catalog."
          }
        ]
      },
      "head": {
        "operationId": "getImageryHead",
        "summary": "Get an exact imagery record (headers only)",
        "tags": [
          "Catalog"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "A valid aarde web token is required"
          },
          "400": {
            "description": "invalid limit"
          },
          "404": {
            "description": "imagery not found"
          },
          "500": {
            "description": "catalog request failed"
          }
        },
        "parameters": [
          {
            "name": "catalogID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Exact, case-sensitive ID: 1\u2013255 UTF-8 bytes, no slashes, backslashes, control characters, or surrounding whitespace."
            },
            "description": "Catalog identifier."
          },
          {
            "name": "imageID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Exact, case-sensitive ID: 1\u2013255 UTF-8 bytes, no slashes, backslashes, control characters, or surrounding whitespace."
            },
            "description": "Image identifier within the catalog."
          }
        ]
      }
    },
    "/api/v1/imagery/search": {
      "post": {
        "operationId": "searchImagery",
        "summary": "Search imagery by polygon intersection",
        "tags": [
          "Catalog"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "400": {
            "$ref": "#/components/responses/InvalidSearch"
          },
          "413": {
            "$ref": "#/components/responses/TooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "description": "Read-only search: never mutates the catalog. Indexed PostGIS ST_Intersects includes partial overlap and boundary contact and respects holes. Results use the same ordering and pagination as GET /imagery. Anonymous requests are allowed when public reads are enabled. Two-dimensional EPSG:4326 coordinates, longitude then latitude. Rings must be closed, with at least four positions each. At most 10,000 positions across the geometry. PostGIS validates topology, including self-intersections. Split antimeridian crossings into a MultiPolygon; edges spanning more than 180 degrees longitude are rejected. GeoJSON Feature and FeatureCollection wrappers are not accepted.",
        "requestBody": {
          "required": true,
          "description": "One JSON object, at most 1 MiB.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Startup web token. Enter the token only, without the Bearer prefix."
      }
    },
    "schemas": {
      "Info": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string"
          },
          "public_read": {
            "type": "boolean"
          },
          "authenticated": {
            "type": "boolean"
          },
          "read_only": {
            "type": "boolean",
            "const": true
          },
          "basemap": {
            "type": "string",
            "enum": [
              "osm",
              "offline",
              "none"
            ]
          }
        },
        "required": [
          "version",
          "public_read",
          "authenticated",
          "read_only",
          "basemap"
        ],
        "example": {
          "version": "dev",
          "public_read": false,
          "authenticated": true,
          "read_only": true,
          "basemap": "osm"
        }
      },
...
      "Catalogs": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Exact, case-sensitive ID: 1\u2013255 UTF-8 bytes, no slashes, backslashes, control characters, or surrounding whitespace."
            }
          }
        },
        "required": [
          "items"
        ],
        "example": {
          "items": [
            "default"
          ]
        }
      },
...
      "SearchResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Imagery"
            }
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200,
            "default": 50
          },
          "offset": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000,
            "default": 0
          },
          "has_more": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "limit",
          "offset",
          "has_more"
        ],
        "example": {
          "items": [],
          "limit": 50,
          "offset": 0,
          "has_more": false
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "A valid aarde web token is required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "unauthorized",
                "message": "A valid aarde web token is required"
              }
            }
          }
        },
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            },
            "example": "Bearer realm=\"aarde-web\""
          }
        }
      },
      "InvalidRequest": {
        "description": "Invalid pagination, catalog/image IDs, RFC3339 date bounds, cloud thresholds, or cloud_cover_unknown policy. Thresholds must be finite percentages from 0 to 100 and cannot be combined. Unknown-only cannot be combined with a threshold. Spatial search requires valid polygon geometry and a single JSON object with known fields.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "invalid_request",
                "message": "invalid limit"
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "imagery not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "not_found",
                "message": "imagery not found"
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "catalog request failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "internal_error",
                "message": "catalog request failed"
              }
            }
          }
        }
      },
...
      "TooLarge": {
        "description": "request body exceeds 1 MiB",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "body_too_large",
                "message": "request body exceeds 1 MiB"
              }
            }
          }
        }
      }
    }
  }
}